Add files via upload
Segger files to flash the Smart Stirrer for inertia sensor. Unzip this into the examples/ble_peripheral/ folder in the nRF SDK.  Copy the stirboard.h file into components/boards folder. Make sure there is an entry for the Stirboard in components/boards/boards.h, it should look like this:
#elif defined(BOARD_STIRBOARD)
#include "stirboard.h"